# openapi.json
{
  "components": {
    "schemas": {
...
      "CaptionsInput": {
        "properties": {
          "assetId": {
            "title": "Assetid",
            "type": "string"
          },
          "clipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clipid"
          },
          "language": {
            "default": "auto",
            "title": "Language",
            "type": "string"
          },
          "rangeMs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RangeMs"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "assetId"
        ],
        "title": "CaptionsInput",
        "type": "object"
      },
...
      "ColorAnalysisInput": {
        "properties": {
          "assetId": {
            "title": "Assetid",
            "type": "string"
          },
          "clipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clipid"
          },
          "preferredLook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferredlook"
          },
          "rangeMs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RangeMs"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "assetId"
        ],
        "title": "ColorAnalysisInput",
        "type": "object"
      },
...
      "CreateJobRequest": {
        "properties": {
          "input": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaptionsInput"
              },
              {
                "$ref": "#/components/schemas/SmartTrimInput"
              },
              {
                "$ref": "#/components/schemas/ColorAnalysisInput"
              },
              {
                "$ref": "#/components/schemas/ExportInput"
              }
            ],
            "title": "Input"
          },
          "projectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projectid"
          },
          "tool": {
            "$ref": "#/components/schemas/JobTool"
          }
        },
        "required": [
          "tool",
          "input"
        ],
        "title": "CreateJobRequest",
        "type": "object"
      },
...
      "EstimateResponse": {
        "description": "What `POST /jobs` would charge, without creating anything.\n\n`blockedBy` carries the error code the real call would return, so the client\ncan put \"Not enough credits\" on the button itself rather than after a failed\nclick (contract \u00a76.1).\n\n`blockedDetails` carries **the same `details` the failed call would have\nsent** \u2014 `requiredPlan`, `required`, `available`. Without it the client can\nsay *that* something is blocked and not *what to do about it*: naming the\nupgrade needs the plan, and a frontend that hardcoded one would have sent a\nfree user to Pro at $19.99 for a 1080p export that `beta` covers at $3.99.\nThe server reads it from the plans table, so it follows a repricing.",
        "properties": {
          "blockedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockedby"
          },
          "blockedDetails": {
            "additionalProperties": true,
            "title": "Blockeddetails",
            "type": "object"
          },
          "credits": {
            "title": "Credits",
            "type": "integer"
          },
          "estimatedSeconds": {
            "title": "Estimatedseconds",
            "type": "integer"
          },
          "sufficientBalance": {
            "title": "Sufficientbalance",
            "type": "boolean"
          },
          "wouldReserveFrom": {
            "$ref": "#/components/schemas/ReservedFrom"
          }
        },
        "required": [
          "credits",
          "wouldReserveFrom",
          "estimatedSeconds",
          "sufficientBalance"
        ],
        "title": "EstimateResponse",
        "type": "object"
      },
      "ExportInput": {
        "description": "The one input with no `assetId`: an export is of a *project*, not of a\nclip. `projectId` on the request body carries it.\n\n**`timelineVersion` is not optimistic locking, it is a refusal to guess.**\nEvery other tool reads an asset that cannot change underneath it. An export\nreads the timeline, and a timeline the user has edited since they pressed\nthe button is a render of something they are no longer looking at \u2014 so a\nmismatch is `409 VERSION_CONFLICT` rather than a render of the current\nversion (contract \u00a76.2).",
        "properties": {
          "preset": {
            "$ref": "#/components/schemas/ExportPreset"
          },
          "timelineVersion": {
            "minimum": 1.0,
            "title": "Timelineversion",
            "type": "integer"
          }
        },
        "required": [
          "timelineVersion"
        ],
        "title": "ExportInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobFamily": {
        "enum": [
          "analysis",
          "render",
          "inference"
        ],
        "title": "JobFamily",
        "type": "string"
      },
      "JobResponse": {
        "properties": {
          "clipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clipid"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "creditsReserved": {
            "title": "Creditsreserved",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "estimatedSeconds": {
            "title": "Estimatedseconds",
            "type": "integer"
          },
          "family": {
            "$ref": "#/components/schemas/JobFamily"
          },
          "finishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "outputAssetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputassetid"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "projectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projectid"
          },
          "reservedFrom": {
            "$ref": "#/components/schemas/ReservedFrom"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "resultUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resulturl"
          },
          "startedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "tool": {
            "$ref": "#/components/schemas/JobTool"
          }
        },
        "required": [
          "id",
          "tool",
          "family",
          "status",
          "progress",
          "priority",
          "creditsReserved",
          "reservedFrom",
          "estimatedSeconds",
          "createdAt"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
      "Page_JobResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[JobResponse]",
        "type": "object"
      },
...
      "SmartTrimInput": {
        "properties": {
          "assetId": {
            "title": "Assetid",
            "type": "string"
          },
          "clipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clipid"
          },
          "rangeMs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RangeMs"
              },
              {
                "type": "null"
              }
            ]
          },
          "strength": {
            "default": "medium",
            "enum": [
              "light",
              "medium",
              "aggressive"
            ],
            "title": "Strength",
            "type": "string"
          }
        },
        "required": [
          "assetId"
        ],
        "title": "SmartTrimInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI Video Editor. The contract for this API is documented in docs/05-api-contract.md; this schema is generated from it and is the source the frontend generates its types from.",
    "title": "ZipZop API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "description": "*\"On reconnect the client calls this to catch up on anything it missed\nwhile disconnected.\"* \u2014 which is why `status=running` is the query that\nmatters, and why the socket is never the source of truth.",
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projectid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The caller's jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "create_job_v1_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start a job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/estimate": {
      "post": {
        "description": "Contract \u00a76.1. Same body as `POST /jobs`, nothing created.\n\nA block is **reported, not raised**: the point of this endpoint is to put\n\"Not enough credits\" on the button instead of after a failed click, so the\nonly failures that raise here are the ones that mean the request itself is\nwrong \u2014 an asset that does not exist, or one that is not ready.",
        "operationId": "estimate_job_v1_jobs_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "What a job would cost, without creating it",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "description": "`409` if it already finished. Cancelling refunds the reservation in full.\n\nThe refund happens here rather than in the worker: a `queued` job has no\nworker to notice, and one that is running checks the status between stages\nand stops without touching credits that have already gone back.",
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel a job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}